Refactor the clean command.

+++ utils.js
@@ -1,15 +1,30 @@
 'use strict';
 
 const fs = require('fs');
+const readline = require('readline');
 
 const OUT = 'out/';
 
+const prompt = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout
+})
+
 function cleanOutput() {
-  fs.readdir(OUT, (e, files) => {
-    files.forEach(file => {
-      fs.unlinkSync(OUT + file);
-    })
-  })
+  return new Promise((resolve, reject) => {
+    let question = "Are you sure? Y or N";
+    prompt.question(question, (answer) => {
+      if (answer === 'Y') {
+        console.log("Cleaning");
+        fs.readdir(OUT, (e, files) => {
+          files.forEach(file => {
+            fs.unlinkSync(OUT + file);
+          })
+        })
+      }
+      resolve();
+    });
+  });
 }
 
 function getRealArguments(args) {
@@ -24,6 +39,7 @@ function getRealArguments(args) {
   return realArgs;
 }
 
+module.exports.OUT = OUT;
 module.exports.cleanOutput = cleanOutput;
 module.exports.getRealArguments = getRealArguments;
-module.exports.OUT = OUT;
+module.exports.prompt = prompt;
